bots/bt-components-showcase-bot/requirements.txt
bubbletea-chat
python-dotenv